requirements.txt
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=10.0.0
tqdm>=4.65.0
PyYAML>=6.0